# docs/api/v1-openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/users": {
      "post": {
        "operationId": "UsersController_create_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Create a new user within the current tenant (companyId)."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Users"
        ]
      },
      "get": {
        "operationId": "UsersController_findAll_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List users for the current tenant (companyId) with optional pagination and filtering."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Users"
        ]
      }
    },
    "/v1/users/{id}": {
      "get": {
        "operationId": "UsersController_findOne_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User ID (UUID)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get a single user by id within the current tenant (companyId)."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Users"
        ]
      },
      "patch": {
        "operationId": "UsersController_update_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User ID (UUID)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Update a single user by id within the current tenant (companyId)."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Users"
        ]
      },
      "delete": {
        "operationId": "UsersController_remove_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User ID (UUID)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Soft delete a single user by id within the current tenant (companyId)."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Users"
        ]
      }
    },
...
    "/v1/sites-projects": {
      "post": {
        "operationId": "SitesProjectsController_create_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSitesProjectDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "SitesProjects"
        ]
      },
      "get": {
        "operationId": "SitesProjectsController_findAll_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "SitesProjects"
        ]
      }
    },
    "/v1/sites-projects/{id}": {
      "get": {
        "operationId": "SitesProjectsController_findOne_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "SitesProjects"
        ]
      },
      "patch": {
        "operationId": "SitesProjectsController_update_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSitesProjectDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "SitesProjects"
        ]
      },
      "delete": {
        "operationId": "SitesProjectsController_remove_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "SitesProjects"
        ]
      }
    },
    "/v1/safety-reports": {
      "post": {
        "operationId": "SafetyReportsController_create_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSafetyReportDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Safety Reports"
        ]
      },
      "get": {
        "operationId": "SafetyReportsController_findAll_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Safety Reports"
        ]
      }
    },
    "/v1/safety-reports/{id}": {
      "get": {
        "operationId": "SafetyReportsController_findOne_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Safety Reports"
        ]
      },
      "patch": {
        "operationId": "SafetyReportsController_update_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSafetyReportDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Safety Reports"
        ]
      },
      "delete": {
        "operationId": "SafetyReportsController_remove_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Safety Reports"
        ]
      }
    },
...
  },
  "info": {
    "title": "Hawana HSE Core API",
    "description": "HSE SaaS Core Backend (Multi-tenant)",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateSitesProjectDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateSitesProjectDto": {
        "type": "object",
        "properties": {}
      },
      "CreateSafetyReportDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateSafetyReportDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}